requirements.txt
astroid==2.3.3
backcall==0.1.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.8
decorator==4.4.1
Django==2.2.7
django-appconf==1.0.3
django-bootstrap4==1.0.1
django-extensions==2.2.5
django-imagekit==4.0.2
Faker==2.0.4
Flask==1.1.1
idna==2.8
ipython==7.10.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
ndg-httpsclient==0.5.1
parso==0.5.1
pickleshare==0.7.5
pilkit==2.0
Pillow==6.2.1
prompt-toolkit==3.0.2
pusher==2.1.4
pyasn1==0.4.8
pycparser==2.19
Pygments==2.5.2
pylint==2.4.3
pylint-django==2.0.12
pylint-plugin-utils==0.6
PyNaCl==1.3.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
